1. 클래스 이름 변경 2. instructMove 메소드 기능 정리
